Em relação ao projéteis, foram adicionados: sprite, controle de direção e detecção de colisão

+++ Projeto 1904/inimigo.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 
 typedef struct {
-	int botX, botY, larguraBot, alturaBot;
+	float botX, botY, larguraBot, alturaBot;
 }Inimigo;
 
 void perseguir(Inimigo* bot, Jogador* jogador, float larguraJogador, float alturaJogador, float velocidadeInimigo);
